Move health and docs routes to meta module
